backend/requirements.txt
## utils ##
retry
python-dotenv
requests
rich
PyYAML

## redis ##
redis

## celery ##
celery
celery[redis]

## qdrant ##
qdrant_client

## llama index ##
llama_index
langchain